feat(syllabus): Add initial migration files for Syllabus model and update user model with program field - Created initial migration for Syllabus model with fields for file, version, and metadata - Added migration to alter Syllabus model options for ordering - Introduced program field in User model to associate users with programs - Added utility functions for image file validation and profile picture path generation
