+++ backend/api/apps/users/models.py
@@ -1,21 +1,7 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
-from django.core.exceptions import ValidationError
-import os
-
-def validate_image_file(value):
-    ext = os.path.splitext(value.name)[1]
-    valid_extensions = ['.jpg', '.jpeg', '.png', '.gif']
-    if not ext.lower() in valid_extensions:
-        raise ValidationError('Unsupported file extension. Please use: .jpg, .jpeg, .png, or .gif')
-    if value.size > 5 * 1024 * 1024:  # 5MB limit
-        raise ValidationError('File size too large. Size should not exceed 5MB.')
-
-def profile_picture_path(instance, filename):
-    # Get the file extension
-    ext = os.path.splitext(filename)[1]
-    # Return the new path
-    return f'profile_pictures/{instance.id}/{instance.username}{ext}'
+from django.core.validators import FileExtensionValidator
+from .utils import profile_picture_path, validate_image_file
 
 class User(AbstractUser):
     email = models.EmailField(unique=True)
@@ -31,6 +17,13 @@ class User(AbstractUser):
     phone_number = models.CharField(max_length=15, blank=True)
     college = models.CharField(max_length=255, blank=True)
     semester = models.IntegerField(null=True, blank=True)
+    program = models.ForeignKey(
+        'question_papers.Program',
+        on_delete=models.SET_NULL,
+        null=True,
+        blank=True,
+        related_name='students'
+    )
     profile_picture = models.ImageField(
         upload_to=profile_picture_path,
         validators=[validate_image_file],
